refactor: update auction schema and remove unused components
- Replace auction schema import path to use the new auction-mock schema.
- Remove unused components: DataTableToolbar, UserNav, and data.tsx.
- Update columns component to reflect the new auction schema import.
- Ensure code clarity and maintainability by cleaning up unnecessary files.
