--- src/features/pre-auction/auction-maintenance/components/to-remove/user-nav.tsx
@@ -1,58 +1,0 @@
-// import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
-// import { Button } from '@/components/ui/button'
-// import {
-//    DropdownMenu,
-//    DropdownMenuContent,
-//    DropdownMenuGroup,
-//    DropdownMenuItem,
-//    DropdownMenuLabel,
-//    DropdownMenuSeparator,
-//    DropdownMenuShortcut,
-//    DropdownMenuTrigger
-// } from '@/components/ui/dropdown-menu'
-
-// export function UserNav() {
-//    return (
-//       <DropdownMenu>
-//          <DropdownMenuTrigger asChild>
-//             <Button variant="ghost" className="relative h-8 w-8 rounded-full">
-//                <Avatar className="h-9 w-9">
-//                   <AvatarImage src="/avatars/03.png" alt="@shadcn" />
-//                   <AvatarFallback>SC</AvatarFallback>
-//                </Avatar>
-//             </Button>
-//          </DropdownMenuTrigger>
-//          <DropdownMenuContent className="w-56" align="end" forceMount>
-//             <DropdownMenuLabel className="font-normal">
-//                <div className="flex flex-col space-y-1">
-//                   <p className="text-sm font-medium leading-none">shadcn</p>
-//                   <p className="text-xs leading-none text-muted-foreground">
-//                      m@example.com
-//                   </p>
-//                </div>
-//             </DropdownMenuLabel>
-//             <DropdownMenuSeparator />
-//             <DropdownMenuGroup>
-//                <DropdownMenuItem>
-//                   Profile
-//                   <DropdownMenuShortcut>⇧⌘P</DropdownMenuShortcut>
-//                </DropdownMenuItem>
-//                <DropdownMenuItem>
-//                   Billing
-//                   <DropdownMenuShortcut>⌘B</DropdownMenuShortcut>
-//                </DropdownMenuItem>
-//                <DropdownMenuItem>
-//                   Settings
-//                   <DropdownMenuShortcut>⌘S</DropdownMenuShortcut>
-//                </DropdownMenuItem>
-//                <DropdownMenuItem>New Team</DropdownMenuItem>
-//             </DropdownMenuGroup>
-//             <DropdownMenuSeparator />
-//             <DropdownMenuItem>
-//                Log out
-//                <DropdownMenuShortcut>⇧⌘Q</DropdownMenuShortcut>
-//             </DropdownMenuItem>
-//          </DropdownMenuContent>
-//       </DropdownMenu>
-//    )
-// }
