--- src/features/pre-auction/auction-maintenance/data/data.tsx
@@ -1,71 +1,0 @@
-import {
-   ArrowDown,
-   ArrowRight,
-   ArrowUp,
-   CheckCircle,
-   Circle,
-   CircleOff,
-   HelpCircle,
-   Timer
-} from 'lucide-react'
-
-export const labels = [
-   {
-      value: 'bug',
-      label: 'Bug'
-   },
-   {
-      value: 'feature',
-      label: 'Feature'
-   },
-   {
-      value: 'documentation',
-      label: 'Documentation'
-   }
-]
-
-export const statuses = [
-   {
-      value: 'backlog',
-      label: 'Backlog',
-      icon: HelpCircle
-   },
-   {
-      value: 'todo',
-      label: 'Todo',
-      icon: Circle
-   },
-   {
-      value: 'in progress',
-      label: 'In Progress',
-      icon: Timer
-   },
-   {
-      value: 'done',
-      label: 'Done',
-      icon: CheckCircle
-   },
-   {
-      value: 'canceled',
-      label: 'Canceled',
-      icon: CircleOff
-   }
-]
-
-export const priorities = [
-   {
-      label: 'Low',
-      value: 'low',
-      icon: ArrowDown
-   },
-   {
-      label: 'Medium',
-      value: 'medium',
-      icon: ArrowRight
-   },
-   {
-      label: 'High',
-      value: 'high',
-      icon: ArrowUp
-   }
-]
