+++ src/features/pre-auction/auction-maintenance/schemas/auction-mock.ts
